Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4288



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@17306 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/gencase/store.c
@@ -494,7 +494,7 @@ makeUnfoldData() {
         for(i=0; i<unfoldRows; ++i) {
             p+=UGENCASE_UNFOLD_WIDTH;
             printf("[%2d] %04x %04x %04x <- %04x %04x\n",
-                   i, p[0], p[1], p[2], p[3], p[4]);
+                   (int)i, p[0], p[1], p[2], p[3], p[4]);
         }
     }
 }
